moved keyboard specific code from keymap.c to separate file. also using *_kb functions instead of *_user
