+++ keyboards/tsc/tiwaz/config.h
@@ -2,5 +2,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #pragma once
-#define SPLIT_TRANSACTION_IDS_USER RPC_JOYSTICK_AXES, RPC_USER_CONFIG
+#define SPLIT_TRANSACTION_IDS_KB RPC_JOYSTICK_AXES, RPC_KB_CONFIG
 #define RGB_MATRIX_SPLIT { 44, 44 }
